+++ CHANGELOG.md
@@ -16,6 +16,9 @@ CHANGELOG
 ### Management
 * Reload and restart have been added as actions on bots and the whole botnet
 
+### Management
+* A click on the bot name opens the monitor page of the bot
+
 0.2.1
 -----
 * Fix syntax error in pipeline.js preventing loading of configuration page